+++ docs/_pages/releases.md
@@ -11,6 +11,7 @@ sidebar:
 
 ### What's new
 
+* Added `Excluding<T>()` and `Excluding(Type)` overloads to exclude all members of a certain type from equivalency comparisons - [#3114](https://github.com/fluentassertions/fluentassertions/pull/3114)
 * New overloads of `BeXmlSerializable` and `BeDataContractSerializable` that allow `EquivalencyOptions` to be configured, allowing the use of member selection rules. - [#3107](https://github.com/fluentassertions/fluentassertions/pull/3107)
 
 ## 8.8.0